Move rules package to genesis
