+++ examples/morpheusvm/tests/integration/integration_test.go
@@ -41,13 +41,13 @@ import (
 	"github.com/ava-labs/hypersdk/fees"
 	"github.com/ava-labs/hypersdk/pubsub"
 	"github.com/ava-labs/hypersdk/rpc"
-	"github.com/ava-labs/hypersdk/rules"
 	"github.com/ava-labs/hypersdk/vm"
 
 	auth "github.com/ava-labs/hypersdk/auth"
 	hbls "github.com/ava-labs/hypersdk/crypto/bls"
 	lconsts "github.com/ava-labs/hypersdk/examples/morpheusvm/consts"
 	lrpc "github.com/ava-labs/hypersdk/examples/morpheusvm/rpc"
+	hgenesis "github.com/ava-labs/hypersdk/genesis"
 	hutils "github.com/ava-labs/hypersdk/utils"
 	ginkgo "github.com/onsi/ginkgo/v2"
 )
@@ -177,7 +177,7 @@ var _ = ginkgo.BeforeSuite(func() {
 	gen = genesis.Default()
 	gen.MinUnitPrice = fees.Dimensions{1, 1, 1, 1, 1}
 	gen.MinBlockGap = 0
-	gen.CustomAllocation = []*rules.CustomAllocation{
+	gen.CustomAllocation = []*hgenesis.CustomAllocation{
 		{
 			Address: addrStr,
 			Balance: 10_000_000,
