+++ genesis/genesis.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package rules
+package genesis
 
 import (
 	"github.com/ava-labs/avalanchego/x/merkledb"
